Major Refactor
I updated the binding to use the knockout subscriptions instead of the beforeRenderAll event callback.  When removing array items, the datatables were being destroyed too late causing the bindings and knockout to act weird. I suspect this is because of some race conditions that were happening... anyways, switching to the subscribe method fixed this.